requirements.txt
# Core ML & Data Science Libraries
numpy==1.26
pandas
scipy
scikit-learn
matplotlib
seaborn

# Cheminformatics & Bioinformatics
rdkit
openmm
mdanalysis
biopython
openmmforcefields
openff-toolkit

# Deep Learning & OCR
# Note: These are large libraries. Install them only if you need OCR/vision features.
# paddlepaddle-gpu
# paddleocr

# LLM & Frameworks
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
openai
anthropic
google-generativeai

# Database & Utilities
neo4j
tqdm
requests
beautifulsoup4
pyopenssl
ax-platform
pubmed-lookup
PyMuPDF
pdfplumber
python-dotenv
crossrefapi
pydantic

# Botorch dependencies for Bayesian Optimization
botorch
gpytorch
pyro-ppl
linear-operator

# Editable install for local development
-e .
